funds-manager: funds-manager-api: Refactor crates for custody API
