+++ Cargo.toml
@@ -4,7 +4,8 @@ members = [
     "compliance/compliance-api",
     "dealer/renegade-dealer",
     "dealer/renegade-dealer-api",
-    "funds-manager",
+    "funds-manager/funds-manager-api",
+    "funds-manager/funds-manager-server",
     "price-reporter",
 ]
