+++ funds-manager/funds-manager-api/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "funds-manager-api"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
