+++ funds-manager/funds-manager-api/src/lib.rs
@@ -1,0 +1,14 @@
+//! The API for the funds manager
+#![deny(missing_docs)]
+#![deny(clippy::missing_docs_in_private_items)]
+
+// --------------
+// | Api Routes |
+// --------------
+
+/// The ping route
+pub const PING_ROUTE: &str = "ping";
+/// The route through which a client may start the fee indexing process
+pub const INDEX_FEES_ROUTE: &str = "index-fees";
+/// The route through which a client may start the fee redemption process
+pub const REDEEM_FEES_ROUTE: &str = "redeem-fees";
