+++ funds-manager/funds-manager-server/Cargo.toml
@@ -7,10 +7,10 @@ edition = "2021"
 [dependencies]
 # === CLI + Server === #
 clap = { version = "4.5.3", features = ["derive", "env"] }
+funds-manager-api = { path = "../funds-manager-api" }
 http-body-util = "0.1.0"
 tokio = { version = "1.10", features = ["full"] }
 warp = "0.3"
-
 
 # === Infra === #
 aws-sdk-secretsmanager = "1.37"
