+++ funds-manager/funds-manager-server/src/error.rs
