+++ funds-manager/funds-manager-server/src/fee_indexer/index_fees.rs
