+++ funds-manager/funds-manager-server/src/fee_indexer/queries.rs
